feat: parse uri without port
This allows for node connection uris without a specified port to be
parsed and assigned a default port of 8885. This allows the port to be
left off in the rpc `Connect` call when connecting to the default port.
