+++ lib/utils/utils.ts
@@ -1,16 +1,9 @@
 import http from 'http';
 import p2pErrors from '../p2p/errors';
-import { assert } from 'chai';
 import { Pair } from '../types/orders';
 import crypto from 'crypto';
 import { promisify } from 'util';
 import moment from 'moment';
-
-export type UriParts = {
-  nodePubKey: string;
-  host: string;
-  port: number;
-};
 
 /**
  * Gets the external IP of the node.
@@ -36,30 +29,6 @@ export const getExternalIp = () => {
       reject(p2pErrors.EXTERNAL_IP_UNRETRIEVABLE(err));
     });
   });
-};
-
-/**
- * Creates a URI from the public key, host and port.
- */
-export const getUri = (uriParts: UriParts): string => {
-  const { nodePubKey, host, port } = uriParts;
-  return `${nodePubKey}@${host}:${port}`;
-};
-
-/**
- * Splits a URI into the public key, host and port.
- */
-export const parseUri = (uri: string): UriParts => {
-  // A regex that splits the string by the symbols "@" and ":"
-  const split = uri.split(/[@:]+/);
-
-  assert(split.length === 3);
-
-  return {
-    nodePubKey: split[0],
-    host: split[1],
-    port: Number(split[2]),
-  };
 };
 
 /**
